operator::req_convert::insert::fill_impure_default

Function fill_reqs_with_impure_default

pub fn fill_reqs_with_impure_default(
    table_infos: &HashMap<TableId, Arc<TableInfo>>,
    inserts: InstantAndNormalInsertRequests,
) -> Result<InstantAndNormalInsertRequests>

Fill impure default values in the request(only for normal insert requests, since instant insert can be filled in flownode directly as a single source of truth)