operator::req_convert::insert::stmt_to_region

Function sql_value_to_value

fn sql_value_to_value(
    column_schema: &ColumnSchema,
    sql_val: &Value,
    timezone: Option<&Timezone>,
    auto_string_to_numeric: bool,
) -> Result<Value>

Converts SQL value to gRPC value according to the column schema. If auto_string_to_numeric is true, tries to cast the string value to numeric values, and fills the default value if the cast fails.