Module operator::statement

Modules

copy_database 🔒

copy_table_from 🔒

copy_table_to 🔒

ddl 🔒

describe 🔒

dml 🔒

show 🔒

tql 🔒

Structs

StatementExecutor

Functions

extract_timestamp 🔒

Extracts timestamp from a [HashMap<String, String>] with given key.

idents_to_full_database_name 🔒

set_bytea_output 🔒

set_timezone 🔒

timestamp_range_from_option_map 🔒

Extracts timestamp range from OptionMap with keys start_time and end_time. The timestamp ranges should be a valid timestamp string as defined in Timestamp::from_str. The timezone used for conversion will respect that inside query_ctx.

to_copy_database_request 🔒

Converts CopyDatabaseArgument to CopyDatabaseRequest. This function extracts the necessary info including catalog/database name, time range, etc.

to_copy_table_request 🔒

validate_client_encoding 🔒