operator::statement

Module admin

Re-exports

pub use layer::AdminEventRecorderHandle;

pub use layer::AdminFunctionLayer;

pub use layer::AdminFunctionLayerRef;

pub use layer::AdminFunctionRecordingLayer;

Modules

event 🔒

layer 🔒

Structs

AdminFunctionRequest

A request to execute one ADMIN function statement.

AdminFunctionResponse

The client output and immediate typed result of one ADMIN function execution.

CoreAdminFunctionService 🔒

Constants

DUMMY_COLUMN 🔒

Traits

AdminFunctionService

Executes an ADMIN function request.

Functions

args_to_vector 🔒

Try to cast the arguments to vectors by function’s signature.

immediate_result 🔒

new_admin_function_service 🔒

Creates the core ADMIN function service.

try_get_data_type_for_sql_value 🔒

Try to infer the data type from sql value.

value_to_vector 🔒

Build a [VectorRef] from [Value]

values_to_vectors_by_exact_types 🔒

Try to cast sql values to vectors by exact data types.

values_to_vectors_by_valid_types 🔒

Try to cast sql values to vectors by valid data types.

Type Aliases

AdminFunctionServiceRef

A shared ADMIN function service.