operator::statement::copy_table_from
Function generated_schema_projection_and_compatible_file_schema

fn generated_schema_projection_and_compatible_file_schema(
    file: &SchemaRef,
    table: &SchemaRef,
) -> (Vec<usize>, Vec<usize>, Schema)

Generates a maybe compatible schema of the file schema.

If there is a field is found in table schema, copy the field data type to maybe compatible schema(compatible_fields).