operator::req_convert::insert

Module stmt_to_region

Structs

StatementToRegion

Constants

DEFAULT_PLACEHOLDER_VALUE 🔒

Functions

column_names 🔒

replace_default 🔒

sql_value_to_grpc_value 🔒

Converts SQL value to gRPC value according to the column schema. If auto_string_to_numeric is true, tries to cast the string value to numeric values, and fills the default value if the cast fails.