operator

Module statement

Modules

admin 🔒

comment 🔒

copy_database 🔒

copy_query_to 🔒

copy_table_from 🔒

copy_table_to 🔒

cursor 🔒

ddl

describe 🔒

dml 🔒

kill 🔒

set 🔒

show 🔒

tql 🔒

Structs

ExecutorConfigureContext

InserterImpl

The [Inserter] implementation for the statement executor.

StatementExecutor

Traits

StatementExecutorConfigurator

A configurator that customizes or enhances a StatementExecutor.

Functions

extract_timestamp 🔒

Extracts timestamp from a HashMap<String, String> with given key.

idents_to_full_database_name 🔒

timestamp_range_from_option_map 🔒

Extracts timestamp range from OptionMap with keys start_time and end_time. The timestamp ranges should be a valid timestamp string as defined in [Timestamp::from_str]. The timezone used for conversion will respect that inside query_ctx.

to_copy_database_request 🔒

Converts [CopyDatabaseArgument] to [CopyDatabaseRequest]. This function extracts the necessary info including catalog/database name, time range, etc.

to_copy_query_request 🔒

to_copy_table_request 🔒

verify_time_related_format 🔒

Type Aliases

StatementExecutorConfiguratorRef

StatementExecutorRef