operator

Module statement

Modules

admin ğŸ”’

copy_database ğŸ”’

copy_query_to ğŸ”’

copy_table_from ğŸ”’

copy_table_to ğŸ”’

cursor ğŸ”’

ddl

describe ğŸ”’

dml ğŸ”’

kill ğŸ”’

set ğŸ”’

show ğŸ”’

tql ğŸ”’

Structs

InserterImpl

The Inserter implementation for the statement executor.

StatementExecutor

Functions

extract_timestamp ğŸ”’

Extracts timestamp from a HashMap<String, String> with given key.

idents_to_full_database_name ğŸ”’

timestamp_range_from_option_map ğŸ”’

Extracts timestamp range from OptionMap with keys start_time and end_time. The timestamp ranges should be a valid timestamp string as defined in Timestamp::from_str. The timezone used for conversion will respect that inside query_ctx.

to_copy_database_request ğŸ”’

Converts CopyDatabaseArgument to CopyDatabaseRequest. This function extracts the necessary info including catalog/database name, time range, etc.

to_copy_query_request ğŸ”’

to_copy_table_request ğŸ”’

Type Aliases

StatementExecutorRef