Module operator::statement

Modules

copy_database ğŸ”’

copy_table_from ğŸ”’

copy_table_to ğŸ”’

ddl ğŸ”’

describe ğŸ”’

dml ğŸ”’

set ğŸ”’

show ğŸ”’

tql ğŸ”’

Structs

StatementExecutor

Functions

extract_timestamp ğŸ”’

Extracts timestamp from a [HashMap<String, String>] with given key.

idents_to_full_database_name ğŸ”’

timestamp_range_from_option_map ğŸ”’

Extracts timestamp range from OptionMap with keys start_time and end_time. The timestamp ranges should be a valid timestamp string as defined in [Timestamp::from_str]. The timezone used for conversion will respect that inside query_ctx.

to_copy_database_request ğŸ”’

Converts [CopyDatabaseArgument] to [CopyDatabaseRequest]. This function extracts the necessary info including catalog/database name, time range, etc.

to_copy_table_request ğŸ”’

Type Aliases

StatementExecutorRef