Module operator::statement::admin

Constants

DUMMY_COLUMN 🔒

Functions

args_to_vector 🔒

Try to cast the arguments to vectors by function’s signature.

try_get_data_type_for_sql_value 🔒

Try to infer the data type from sql value.

value_to_vector 🔒

Build a [VectorRef] from [Value]

values_to_vectors_by_exact_types 🔒

Try to cast sql values to vectors by exact data types.

values_to_vectors_by_valid_types 🔒

Try to cast sql values to vectors by valid data types.